requirements.txt
aiogram==3.13.1
requests==2.32.3
aioschedule==0.3.1
python-dotenv==1.0.1
